encoding_ipfs_done/requirements.txt
opencv-python
requests
pycryptodome
cryptography
numpy